Create Tools123.sh
I added functionality that prompts the user to send the report via email. If the user agrees, the program generates an HTML file in a specific directory in Kali Linux. When the user opens the file, it launches a web browser. By clicking on the link in the HTML file, the user can open an email service like Gmail or Yahoo. From there, the user can send the report to anyone.